# ARROW-10765: [Rust] Optimize take string for non-null arrays