--- a/rust/arrow/src/compute/kernels/take.rs
+++ b/rust/arrow/src/compute/kernels/take.rs
@@ -365,43 +365,81 @@ where
         .unwrap();
 
     let num_bytes = bit_util::ceil(data_len, 8);
-    let mut null_buf = MutableBuffer::new(num_bytes).with_bitset(num_bytes, true);
-    let null_slice = null_buf.data_mut();
 
+    let null_count = array.null_count();
     let mut offsets = Vec::with_capacity(data_len + 1);
     let mut values = Vec::with_capacity(data_len);
     let mut length_so_far = OffsetSize::zero();
 
-    offsets.push(length_so_far);
-    for i in 0..data_len {
-        let index = ToPrimitive::to_usize(&indices.value(i)).ok_or_else(|| {
-            ArrowError::ComputeError("Cast to usize failed".to_string())
-        })?;
+    let nulls;
+    if null_count == 0 && indices.null_count() == 0 {
+        for i in 0..data_len {
+            let index = ToPrimitive::to_usize(&indices.value(i)).ok_or_else(|| {
+                ArrowError::ComputeError("Cast to usize failed".to_string())
+            })?;
 
-        if array.is_valid(index) && indices.is_valid(i) {
             let s = array.value(index);
 
             length_so_far += OffsetSize::from_usize(s.len()).unwrap();
             values.extend_from_slice(s.as_bytes());
-        } else {
-            // set null bit
-            bit_util::unset_bit(null_slice, i);
+            offsets.push(length_so_far);
         }
+        nulls = None
+    } else if null_count == 0 {
+        for i in 0..data_len {
+            let index = ToPrimitive::to_usize(&indices.value(i)).ok_or_else(|| {
+                ArrowError::ComputeError("Cast to usize failed".to_string())
+            })?;
+
+            if indices.is_valid(i) {
+                let s = array.value(index);
+
+                length_so_far += OffsetSize::from_usize(s.len()).unwrap();
+                values.extend_from_slice(s.as_bytes());
+            }
+            offsets.push(length_so_far);
+        }
+        nulls = None
+    } else {
+        let mut null_buf = MutableBuffer::new(num_bytes).with_bitset(num_bytes, true);
+        let null_slice = null_buf.data_mut();
+
         offsets.push(length_so_far);
+        for i in 0..data_len {
+            let index = ToPrimitive::to_usize(&indices.value(i)).ok_or_else(|| {
+                ArrowError::ComputeError("Cast to usize failed".to_string())
+            })?;
+
+            if array.is_valid(index) && indices.is_valid(i) {
+                let s = array.value(index);
+
+                length_so_far += OffsetSize::from_usize(s.len()).unwrap();
+                values.extend_from_slice(s.as_bytes());
+            } else {
+                // set null bit
+                bit_util::unset_bit(null_slice, i);
+            }
+            offsets.push(length_so_far);
+        }
+
+        nulls = match indices.data_ref().null_buffer() {
+            Some(buffer) => {
+                Some(buffer_bin_and(buffer, 0, &null_buf.freeze(), 0, data_len))
+            }
+            None => Some(null_buf.freeze()),
+        };
     }
 
-    let nulls = match indices.data_ref().null_buffer() {
-        Some(buffer) => buffer_bin_and(buffer, 0, &null_buf.freeze(), 0, data_len),
-        None => null_buf.freeze(),
-    };
-
-    let data = ArrayData::builder(<OffsetSize as StringOffsetSizeTrait>::DATA_TYPE)
+    let mut data = ArrayData::builder(<OffsetSize as StringOffsetSizeTrait>::DATA_TYPE)
         .len(data_len)
-        .null_bit_buffer(nulls)
         .add_buffer(Buffer::from(offsets.to_byte_slice()))
-        .add_buffer(Buffer::from(&values[..]))
-        .build();
-    Ok(Arc::new(GenericStringArray::<OffsetSize>::from(data)))
+        .add_buffer(Buffer::from(&values[..]));
+    if let Some(null_buffer) = nulls {
+        data = data.add_buffer(null_buffer);
+    }
+    Ok(Arc::new(GenericStringArray::<OffsetSize>::from(
+        data.build(),
+    )))
 }
 
 /// `take` implementation for list arrays