--- a/rust/arrow/src/compute/kernels/take.rs
+++ b/rust/arrow/src/compute/kernels/take.rs
@@ -388,11 +388,11 @@ where
         nulls = None
     } else if null_count == 0 {
         for i in 0..data_len {
-            let index = ToPrimitive::to_usize(&indices.value(i)).ok_or_else(|| {
-                ArrowError::ComputeError("Cast to usize failed".to_string())
-            })?;
-
             if indices.is_valid(i) {
+                let index = ToPrimitive::to_usize(&indices.value(i)).ok_or_else(|| {
+                    ArrowError::ComputeError("Cast to usize failed".to_string())
+                })?;
+    
                 let s = array.value(index);
 
                 length_so_far += OffsetSize::from_usize(s.len()).unwrap();